move tests into subfolder
