+++ test/pipeable.js
@@ -1,33 +1,33 @@
 import test from 'ava'
 import sinon from 'sinon'
 
-import { from } from './lib'
-import { append } from './lib/pipeable/append'
-import { concat } from './lib/pipeable/concat'
-import { first } from './lib/pipeable/first'
-import { eachChunk } from './lib/pipeable/eachChunk'
-import { filter } from './lib/pipeable/filter'
-import { forEach } from './lib/pipeable/forEach'
-import { last } from './lib/pipeable/last'
-import { map } from './lib/pipeable/map'
-import { prepend } from './lib/pipeable/prepend'
-import { reduce } from './lib/pipeable/reduce'
-import { reject } from './lib/pipeable/reject'
-import { scan } from './lib/pipeable/scan'
-import { seq } from './lib/pipeable/seq'
-import { skip } from './lib/pipeable/skip'
-import { skipUntil } from './lib/pipeable/skipUntil'
-import { skipWhile } from './lib/pipeable/skipWhile'
-import { take } from './lib/pipeable/take'
-import { takeUntil } from './lib/pipeable/takeUntil'
-import { takeWhile } from './lib/pipeable/takeWhile'
-import { zip } from './lib/pipeable/zip'
-import { zipAll } from './lib/pipeable/zipAll'
-import { zipWith } from './lib/pipeable/zipWith'
+import { from } from '../commonjs'
+import { append } from '../commonjs/pipeable'
+import { concat } from '../commonjs/pipeable'
+import { first } from '../commonjs/pipeable'
+import { eachChunk } from '../commonjs/pipeable'
+import { filter } from '../commonjs/pipeable'
+import { forEach } from '../commonjs/pipeable'
+import { last } from '../commonjs/pipeable'
+import { map } from '../commonjs/pipeable'
+import { prepend } from '../commonjs/pipeable'
+import { reduce } from '../commonjs/pipeable'
+import { reject } from '../commonjs/pipeable'
+import { scan } from '../commonjs/pipeable'
+import { seq } from '../commonjs/pipeable'
+import { skip } from '../commonjs/pipeable'
+import { skipUntil } from '../commonjs/pipeable'
+import { skipWhile } from '../commonjs/pipeable'
+import { take } from '../commonjs/pipeable'
+import { takeUntil } from '../commonjs/pipeable'
+import { takeWhile } from '../commonjs/pipeable'
+import { zip } from '../commonjs/pipeable'
+import { zipAll } from '../commonjs/pipeable'
+import { zipWith } from '../commonjs/pipeable'
 
-import { toArray } from './lib/pipeable/toArray'
-import { toMap } from './lib/pipeable/toMap'
-import { toSet } from './lib/pipeable/toSet'
+import { toArray } from '../commonjs/pipeable'
+import { toMap } from '../commonjs/pipeable'
+import { toSet } from '../commonjs/pipeable'
 
 const arrayStrings = ['a', 'b', 'c', 'd']
 const arrayNums = [1, 2, 3, 4]
